Undo moved function.
