+++ test/util/test_tools.hpp
@@ -72,6 +72,8 @@ std::vector<T> read_vector(std::shared_ptr<ngraph::runtime::Tensor> tv)
     return rc;
 }
 
+std::vector<float> read_float_vector(std::shared_ptr<ngraph::runtime::Tensor> tv);
+
 template <typename T>
 void write_vector(std::shared_ptr<ngraph::runtime::Tensor> tv, const std::vector<T>& values)
 {
@@ -250,6 +252,38 @@ template <>
 std::string get_results_str(const std::vector<char>& ref_data,
                             const std::vector<char>& actual_data,
                             size_t max_results);
+
+/// \brief      Reads a binary file to a vector.
+///
+/// \param[in]  path  The path where the file is located.
+///
+/// \tparam     T     The type we want to interpret as the elements in binary file.
+///
+/// \return     Return vector of data read from input binary file.
+///
+template <typename T>
+std::vector<T> read_binary_file(const std::string& path)
+{
+    std::vector<T> file_content;
+    std::ifstream inputs_fs{path, std::ios::in | std::ios::binary};
+    if (!inputs_fs)
+    {
+        throw std::runtime_error("Failed to open the file: " + path);
+    }
+
+    inputs_fs.seekg(0, std::ios::end);
+    auto size = inputs_fs.tellg();
+    inputs_fs.seekg(0, std::ios::beg);
+    if (size % sizeof(T) != 0)
+    {
+        throw std::runtime_error(
+            "Error reading binary file content: Input file size (in bytes) "
+            "is not a multiple of requested data type size.");
+    }
+    file_content.resize(size / sizeof(T));
+    inputs_fs.read(reinterpret_cast<char*>(file_content.data()), size);
+    return file_content;
+}
 
 testing::AssertionResult test_ordered_ops(std::shared_ptr<ngraph::Function> f,
                                           const ngraph::NodeVector& required_ops);
